refactor(SPV-1061): arc requests moved to a single package
