+++ engine/chain/interface.go
@@ -6,18 +6,13 @@ import (
 	"github.com/bitcoin-sv/spv-wallet/engine/chain/models"
 )
 
-// QueryService for querying transactions.
-type QueryService interface {
+// ARCService for querying ARC server.
+type ARCService interface {
 	QueryTransaction(ctx context.Context, txID string) (*chainmodels.TXInfo, error)
-}
-
-// PolicyService for querying policy.
-type PolicyService interface {
 	GetPolicy(ctx context.Context) (*chainmodels.Policy, error)
 }
 
 // Service related to the chain.
 type Service interface {
-	QueryService
-	PolicyService
+	ARCService
 }
